added layout component
