+++ pages/_app.tsx
@@ -1,8 +1,13 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
+import { Layout } from '../components/Layout';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
 }
 
 export default MyApp;
