--- src/pages/index.tsx
@@ -1,18 +1,0 @@
-import Head from 'next/head';
-import Link from 'next/link';
-
-export default function Home() {
-  return (
-    <div>
-      <Head>
-        <title>Next | Stripe</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
-        <h1 className="text-3xl">Home Page</h1>
-        <Link href="/products">Go to products page</Link>
-      </main>
-    </div>
-  );
-}
